requirements.txt
requests
bs4
Pillow
tqdm
ShopifyAPI